% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R, R/get_isochrone.R
\name{data_raw_dir}
\alias{data_raw_dir}
\title{data_raw_dir}
\usage{
data_raw_dir()

data_raw_dir()
}
\description{
This function creates a data folder if one doesn't already exist.

This function creates a data folder if one doesn't already exist.
}
\examples{
\dontrun{
}
\dontrun{
}
}
