% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isochrone.R
\name{get_isochrone}
\alias{get_isochrone}
\title{get_isochrone}
\usage{
get_isochrone()
}
\arguments{
\item{src}{list of direct URL to dataset - dataset name to be normalised
with 3 letters country code plus year}
}
\description{
Get all country specific dataset
}
\examples{
\dontrun{
}
}
